moved the remaining proto definitions
