+++ libsql-server/src/hrana/http/mod.rs
@@ -1,6 +1,7 @@
 use anyhow::{Context, Result};
 use bytes::Bytes;
 use futures::stream::Stream;
+use libsql_sys::hrana::proto;
 use parking_lot::Mutex;
 use serde::{de::DeserializeOwned, Serialize};
 use std::pin::Pin;
@@ -12,8 +13,6 @@ use crate::auth::Authenticated;
 use crate::connection::{Connection, MakeConnection};
 use crate::hrana::http::stream::StreamError;
 
-mod proto;
-mod protobuf;
 mod request;
 pub(crate) mod stream;
