+++ libsql/src/hrana/transaction.rs
@@ -1,9 +1,9 @@
-use crate::hrana::pipeline::{ExecuteStreamReq, StreamRequest};
 use crate::hrana::proto::Stmt;
 use crate::hrana::stream::HranaStream;
 use crate::hrana::{HttpSend, Result};
 use crate::parser::StmtKind;
 use crate::TransactionBehavior;
+use libsql_sys::hrana::proto::{ExecuteStreamReq, StreamRequest};
 
 #[derive(Debug, Clone)]
 pub(crate) struct HttpTransaction<T>
