+++ libsql/src/lib.rs
@@ -31,7 +31,7 @@
 //! use libsql::{Database};
 //! use libsql::replication::Frames;
 //!
-//! let mut db = Database::open_with_local_sync("/tmp/test.db").await.unwrap();
+//! let mut db = Database::open_with_local_sync("/tmp/test.db", None).await.unwrap();
 //!
 //! let frames = Frames::Vec(vec![]);
 //! db.sync_frames(frames).await.unwrap();
